// src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
// ...
      vehicles: {
        Row: {
          created_at: string
          description: string | null
          fuel_type: string | null
          id: string
          images: string[] | null
          is_featured: boolean | null
          latitude: number | null
          location: string | null
          longitude: number | null
          make: string
          mileage: number | null
          model: string
          price: number
          seller_id: string | null
          status: string | null
          transmission: string | null
          vin: string | null
          year: number
        }
        Insert: {
          created_at?: string
          description?: string | null
          fuel_type?: string | null
          id?: string
          images?: string[] | null
          is_featured?: boolean | null
          latitude?: number | null
          location?: string | null
          longitude?: number | null
          make: string
          mileage?: number | null
          model: string
          price: number
          seller_id?: string | null
          status?: string | null
          transmission?: string | null
          vin?: string | null
          year: number
        }
        Update: {
          created_at?: string
          description?: string | null
          fuel_type?: string | null
          id?: string
          images?: string[] | null
          is_featured?: boolean | null
          latitude?: number | null
          location?: string | null
          longitude?: number | null
          make?: string
          mileage?: number | null
          model?: string
          price?: number
          seller_id?: string | null
          status?: string | null
          transmission?: string | null
          vin?: string | null
          year?: number
        }
        Relationships: [
          {
            foreignKeyName: "vehicles_seller_id_fkey"
            columns: ["seller_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never